+++ internal/e2e/mirrorregistry/internal/e2etest/registry.go
